requirements.txt
# Core dependencies
python-dotenv==1.0.0
apscheduler==3.10.1
requests==2.31.0
pymongo==4.5.0
redis==4.6.0
loguru==0.7.0
python-telegram-bot==20.3

# Development dependencies
pytest==7.3.1
black==23.3.0
isort==5.12.0
flake8==6.0.0